SLING-9521 Packages exported in earlier API Regions are not available to later API Regions
Added a property to the feature->regions map (features.properties) to
declare the global region order: __region.order__
The value format of __region.order__ is the same as the other entries in
the file: an ordered list of regions. This property lists the total
ordering of regions (while features might only have a partial list).
Since the values of the features.properties have significant ordering, I
also made this explicity my changing the signature of the
feature->regions map to have a LinkedHashSet for the values.